services/auth_service/requirements.txt
asyncpg>=0.31.0
bcrypt>=5.0.0
fastapi>=0.128.0
pydantic-settings>=2.12.0
pydantic[email]>=2.12.5
python-jose[cryptography]>=3.5.0
python-multipart>=0.0.22
sqlalchemy>=2.0.46
uvicorn>=0.40.0